fix: :bug: resolve backup cronjob failure
Logs indicate that the 'packer' user has been executing the
generate_backup.sh script on instances. This is due to the packer
pipeline not executing the crontab command as root, or any other user
that has access to the necessary files/folders to backup.

This change adds the sudo directive to crontab ensuring that the cronjob
is executed as root.